refactor: convert median two arrays to ts
